requirements.txt
python-dotenv
numpy
pandas
scikit-learn
matplotlib
seaborn
pymongo==3.10
certifi
ipykernel
pyaml
dill
#-e .